Moved full cart button functionality into the product interaction component
